+++ Code/Tools/Libs/GuiFoundation/UIServices/Implementation/UIServices.cpp
@@ -14,10 +14,6 @@
 #include <QSettings>
 #include <QUrl>
 
-#if EZ_ENABLED(EZ_PLATFORM_WINDOWS_DESKTOP)
-#  include <ShlObj_core.h>
-#endif
-
 EZ_IMPLEMENT_SINGLETON(ezQtUiServices);
 
 ezEvent<const ezQtUiServices::Event&> ezQtUiServices::s_Events;
@@ -363,93 +359,4 @@ void ezQtUiServices::ShowGlobalStatusBarMessage(const ezFormatString& msg)
 bool ezQtUiServices::OpenFileInDefaultProgram(const char* szPath)
 {
   return QDesktopServices::openUrl(QUrl::fromLocalFile(szPath));
-}
-
-void ezQtUiServices::OpenInExplorer(const char* szPath, bool bIsFile)
-{
-  QStringList args;
-
-#if EZ_ENABLED(EZ_PLATFORM_WINDOWS_DESKTOP)
-  if (bIsFile)
-    args << "/select,";
-
-  args << QDir::toNativeSeparators(szPath);
-
-  QProcess::startDetached("explorer", args);
-#elif EZ_ENABLED(EZ_PLATFORM_LINUX)
-  ezStringBuilder parentDir;
-
-  if (bIsFile)
-  {
-    parentDir = szPath;
-    parentDir = parentDir.GetFileDirectory();
-    szPath = parentDir.GetData();
-  }
-  args << QDir::toNativeSeparators(szPath);
-
-  QProcess::startDetached("xdg-open", args);
-#else
-  EZ_ASSERT_NOT_IMPLEMENTED
-#endif
-}
-
-void ezQtUiServices::OpenWith(const char* szPath)
-{
-  ezStringBuilder sPath = szPath;
-  sPath.MakeCleanPath();
-  sPath.MakePathSeparatorsNative();
-
-#if EZ_ENABLED(EZ_PLATFORM_WINDOWS_DESKTOP)
-  ezStringWChar wpath(sPath);
-  OPENASINFO oi;
-  oi.pcszFile = wpath.GetData();
-  oi.pcszClass = NULL;
-  oi.oaifInFlags = OAIF_EXEC;
-  SHOpenWithDialog(NULL, &oi);
-#else
-  EZ_ASSERT_NOT_IMPLEMENTED
-#endif
-}
-
-ezStatus ezQtUiServices::OpenInVsCode(const QStringList& arguments)
-{
-  QString sVsCodeExe;
-#if EZ_ENABLED(EZ_PLATFORM_WINDOWS)
-  sVsCodeExe =
-    QStandardPaths::locate(QStandardPaths::GenericDataLocation, "Programs/Microsoft VS Code/Code.exe", QStandardPaths::LocateOption::LocateFile);
-
-  if (!QFile().exists(sVsCodeExe))
-  {
-    QSettings settings("\\HKEY_LOCAL_MACHINE\\SOFTWARE\\Classes\\Applications\\Code.exe\\shell\\open\\command", QSettings::NativeFormat);
-    QString sVsCodeExeKey = settings.value(".", "").value<QString>();
-
-    if (sVsCodeExeKey.length() > 5)
-    {
-      // Remove shell parameter and normalize QT Compatible path, QFile expects the file separator to be '/' regardless of operating system
-      sVsCodeExe = sVsCodeExeKey.left(sVsCodeExeKey.length() - 5).replace("\\", "/").replace("\"", "");
-    }
-  }
-#endif
-
-  if (sVsCodeExe.isEmpty() || !QFile().exists(sVsCodeExe))
-  {
-    // Try code executable in PATH
-    if (QProcess::execute("code", {"--version"}) == 0)
-    {
-      sVsCodeExe = "code";
-    }
-    else
-    {
-      return ezStatus("Installation of Visual Studio Code could not be located.\n"
-                      "Please visit 'https://code.visualstudio.com/download' to download the 'User Installer' of Visual Studio Code.");
-    }
-  }
-
-  QProcess proc;
-  if (proc.startDetached(sVsCodeExe, arguments) == false)
-  {
-    return ezStatus("Failed to launch Visual Studio Code.");
-  }
-
-  return ezStatus(EZ_SUCCESS);
 }
